Backend: Move LimaKubernetesBackend to new file.
This splits the kubernetes backend out from the VMBackend file, so that it
will be easier to replace.
