+++ src/backend/factory.ts
@@ -1,6 +1,7 @@
 import os from 'os';
 
 import { Architecture, VMBackend } from './backend';
+import LimaKubernetesBackend from './kube/lima';
 import LimaBackend from './lima';
 import MockBackend from './mock';
 import WSLBackend from './wsl';
@@ -16,9 +17,10 @@ export default function factory(arch: Architecture, dockerDirManager: DockerDirM
 
   switch (platform) {
   case 'linux':
-    return new LimaBackend(arch, dockerDirManager);
   case 'darwin':
-    return new LimaBackend(arch, dockerDirManager);
+    return new LimaBackend(arch, dockerDirManager, (backend: LimaBackend) => {
+      return new LimaKubernetesBackend(arch, backend);
+    });
   case 'win32':
     return new WSLBackend();
   default:
